Renamed files to repersent proper function
