+++ app/src/main/java/com/example/myapplication/HelpPage.java
@@ -1,0 +1,86 @@
+package com.example.myapplication;
+import android.os.Bundle;
+import androidx.fragment.app.Fragment;
+
+import android.animation.ObjectAnimator;
+
+import androidx.annotation.NonNull;
+import androidx.navigation.fragment.NavHostFragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.myapplication.databinding.FragmentHelpBinding;
+
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link HelpPage#newInstance} factory method to * create an instance of this fragment.
+ */
+public class HelpPage extends Fragment {
+
+    // TODO: Rename parameter arguments, choose names that match
+    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+    private static final String ARG_PARAM1 = "param1";
+    private static final String ARG_PARAM2 = "param2";
+
+    // TODO: Rename and change types of parameters
+    private String mParam1;
+    private String mParam2;
+    private ViewGroup v;
+    private FragmentHelpBinding binding;
+    public HelpPage() {
+        // Required empty public constructor
+    }
+
+    /**
+     * Use this factory method to create a new instance of
+     * this fragment using the provided parameters.
+     *
+     * @param param1 Parameter 1.
+     * @param param2 Parameter 2.
+     * @return A new instance of fragment Teleop.
+     */
+    // TODO: Rename and change types and number of parameters
+    public static HelpPage newInstance(String param1, String param2) {
+        HelpPage fragment = new HelpPage();
+        Bundle args = new Bundle();
+        args.putString(ARG_PARAM1, param1);
+        args.putString(ARG_PARAM2, param2);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        if (getArguments() != null) {
+            mParam1 = getArguments().getString(ARG_PARAM1);
+            mParam2 = getArguments().getString(ARG_PARAM2);
+        }
+    }
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        binding = FragmentHelpBinding.inflate(inflater, container, false);
+        v = container;
+        return binding.getRoot();
+    }
+    @Override
+    public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        binding.prevInfoBack.setOnClickListener(v -> UserModel.getMatchData().setBroke(true));
+        ObjectAnimator animation = ObjectAnimator.ofFloat(binding.WolfLogoInfo, "rotation", 0f, 90f, 180f, 270f, 360f, 90f, 180f, 270f, 360f, 90f, 180f, 270f, 360f);
+        animation.setDuration(1000);
+        binding.WolfLogoInfo.setOnClickListener(view1 -> {
+            animation.start();
+            UIHelpers.darkMode = !UIHelpers.darkMode;
+            UIHelpers.lightDark(v, UIHelpers.darkMode);
+        });
+        binding.prevInfoBack.setOnClickListener(view12 -> NavHostFragment.findNavController(HelpPage.this)
+                .navigate(R.id.action_HomePage_to_FirstFragment));
+        UIHelpers.lightDark(v, UIHelpers.darkMode);
+    }
+}
