+++ app/src/main/java/com/example/myapplication/HomePage.java
@@ -1,86 +1,138 @@
 package com.example.myapplication;
-import android.os.Bundle;
-import androidx.fragment.app.Fragment;
-
 import android.animation.ObjectAnimator;
-
-import androidx.annotation.NonNull;
-import androidx.navigation.fragment.NavHostFragment;
-
+import android.app.Activity;
+import android.os.Bundle;
+import android.text.Editable;
+import android.text.TextWatcher;
+import android.util.DisplayMetrics;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.navigation.fragment.NavHostFragment;
 
 import com.example.myapplication.databinding.FragmentHomepageBinding;
-
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link HomePage#newInstance} factory method to * create an instance of this fragment.
- */
+import com.google.android.material.snackbar.Snackbar;
 public class HomePage extends Fragment {
-
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
-    private ViewGroup v;
     private FragmentHomepageBinding binding;
-    public HomePage() {
-        // Required empty public constructor
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment Teleop.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static HomePage newInstance(String param1, String param2) {
-        HomePage fragment = new HomePage();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
+    ViewGroup v;
     @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-    }
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
+    public View onCreateView(
+            @NonNull LayoutInflater  inflater, ViewGroup container,
+            Bundle savedInstanceState
+    ) {
+        ViewModelProvider viewModelProvider = new ViewModelProvider(requireActivity());
+        UserModel userModel = viewModelProvider.get(UserModel.class);
+        MatchData matchData = new MatchData();
+        userModel.setMatchData(matchData);
         binding = FragmentHomepageBinding.inflate(inflater, container, false);
         v = container;
+        String currentTeamNumber = UserModel.getMatchData().getTeamNumber();
+        binding.input.setText(currentTeamNumber);
+
+        String currentMatchNumber = UserModel.getMatchData().getTeamNumber();
+        binding.input.setText(currentMatchNumber);
+
         return binding.getRoot();
     }
-    @Override
+
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        binding.prevInfoBack.setOnClickListener(v -> UserModel.getMatchData().setBroke(true));
-        ObjectAnimator animation = ObjectAnimator.ofFloat(binding.WolfLogoInfo, "rotation", 0f, 90f, 180f, 270f, 360f, 90f, 180f, 270f, 360f, 90f, 180f, 270f, 360f);
+                binding.cont.setOnClickListener(v -> {
+            String teamNumber = String.valueOf(binding.input.getText());
+            String matchNumber = String.valueOf(binding.matchInput.getText());
+            boolean teamNumberCheck = (teamNumber.length() > 0 && teamNumber.length() < 5);
+            boolean matchNumCheck = matchNumber.length() > 0;
+            if (teamNumberCheck && matchNumCheck) {
+                NavHostFragment.findNavController(HomePage.this)
+                        .navigate(R.id.action_FirstFragment_to_ThirdFragment);
+            } else {
+                if (!teamNumberCheck){
+                    Snackbar.make(view, "Invalid team number", 600).show();
+                }
+                if (!matchNumCheck) {
+                    Snackbar.make(view, "Invalid match number", 600).show();
+                }
+
+            }
+        });
+        binding.back.setOnClickListener(v -> NavHostFragment.findNavController(HomePage.this)
+                .navigate(R.id.action_FirstFragment_to_HomePage));
+        ObjectAnimator animation = ObjectAnimator.ofFloat(binding.pop, "rotation", 0f, 90f, 180f, 270f, 360f, 90f, 180f, 270f, 360f, 90f, 180f, 270f, 360f);
         animation.setDuration(1000);
-        binding.WolfLogoInfo.setOnClickListener(view1 -> {
+        binding.pop.setOnClickListener(view1 -> {
             animation.start();
             UIHelpers.darkMode = !UIHelpers.darkMode;
             UIHelpers.lightDark(v, UIHelpers.darkMode);
         });
-        binding.prevInfoBack.setOnClickListener(view12 -> NavHostFragment.findNavController(HomePage.this)
-                .navigate(R.id.action_HomePage_to_FirstFragment));
+        DisplayMetrics dm = new DisplayMetrics();
+        ((Activity) requireContext()).getWindowManager().getDefaultDisplay().getMetrics(dm);
+        float width = dm.widthPixels;
+        float height = dm.heightPixels;
+        binding.title.setTranslationX(height * 0.072f);
+        binding.title.setTranslationX(width * 0.146f);
+        binding.back.setTranslationY(height * 0.270f);
+        binding.back.setTranslationX(width * 0.024f);
+        binding.input.setTranslationY(height * 0.158f);
+
+        binding.input.addTextChangedListener(new TextWatcher() {
+            @Override
+            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
+            }
+
+            @Override
+            public void onTextChanged(CharSequence s, int start, int before, int count) {
+
+            }
+
+            @Override
+            public void afterTextChanged(Editable s) {
+                Editable input = binding.input.getText();
+                if (input == null){
+                    return;
+                }
+                String teamNumber = String.valueOf(input);
+                UserModel.getMatchData().setTeamNumber(teamNumber);
+            }
+        });
+
+        binding.matchInput.addTextChangedListener(new TextWatcher() {
+            @Override
+            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
+            }
+
+            @Override
+            public void onTextChanged(CharSequence s, int start, int before, int count) {
+
+            }
+
+            @Override
+            public void afterTextChanged(Editable s) {
+                Editable input = binding.matchInput.getText();
+                if (input == null){
+                    return;
+                }
+                String matchNumber = String.valueOf(input);
+                UserModel.getMatchData().setMatchNumber(matchNumber);
+            }
+        });
+        binding.cont.setTranslationY(height * 0.270f);
+        binding.cont.setTranslationX(width * 0.707f);
+        binding.pop.setTranslationY(height * 0.719f);
+        binding.pop.setTranslationX(width * 0.073f);
         UIHelpers.lightDark(v, UIHelpers.darkMode);
     }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+
+
 }
+
+
+
