+++ app/src/main/res/layout/fragment_auto.xml
