--- app/src/main/res/layout/fragment_first.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.core.widget.NestedScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".FirstFragment">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="730dp"
-        android:padding="0dp">
-
-        <Button
-            android:id="@+id/cont"
-            android:layout_width="110dp"
-            android:layout_height="wrap_content"
-            android:text="Continue"
-            android:translationY="118dp"
-            android:translationX="290dp"/>
-        <Button
-            android:id="@+id/back"
-            android:layout_width="110dp"
-            android:layout_height="wrap_content"
-            android:text="Back"
-            android:translationY="118dp"
-            android:translationX="10dp"/>
-
-        <TextView
-            android:id="@+id/title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Team Selection"
-            android:textSize="42dp"
-            android:translationX="60dp"
-            android:translationY="50dp"/>
-
-
-        <ImageView
-            android:id="@+id/pop"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/pop"
-            android:translationY="500dp"
-            android:translationX="30dp"/>
-
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/input"
-            android:layout_width="140dp"
-            android:layout_height="66dp"
-            android:hint="Team Number"
-            android:maxLength="4"
-            android:inputType="number"
-            android:translationX="133dp"
-            android:translationY="110dp" />
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/matchInput"
-            android:layout_width="140dp"
-            android:layout_height="66dp"
-            android:hint="Match Number"
-            android:maxLength="3"
-            android:inputType="number"
-            android:translationX="133dp"
-            android:translationY="200dp" />
-
-    </RelativeLayout>
-</androidx.core.widget.NestedScrollView>
