+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,9 +7,9 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.example.myapplication.FirstFragment"
+        android:name="com.example.myapplication.HomePage"
         android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
+        tools:layout="@layout/fragment_homepage">
 
         <action
             android:id="@+id/action_FirstFragment_to_ThirdFragment"
@@ -20,9 +20,9 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.example.myapplication.SecondFragment"
+        android:name="com.example.myapplication.TeleOpPage"
         android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
+        tools:layout="@layout/fragment_teleop">
         <action
             android:id="@+id/action_SecondFragment_to_SecondFragment2"
             app:destination="@id/SecondFragment2" />
@@ -32,9 +32,9 @@
     </fragment>
     <fragment
         android:id="@+id/ThirdFragment"
-        android:name="com.example.myapplication.ThirdFragment"
+        android:name="com.example.myapplication.AutoPage"
         android:label="Autonomous"
-        tools:layout="@layout/fragment_third">
+        tools:layout="@layout/fragment_auto">
 
         <action
             android:id="@+id/action_ThirdFragment_to_FirstFragment"
@@ -46,18 +46,18 @@
 
     <fragment
         android:id="@+id/HomePage"
-        android:name="com.example.myapplication.HomePage"
+        android:name="com.example.myapplication.HelpPage"
         android:label="fragment_homepage"
-        tools:layout="@layout/fragment_homepage" >
+        tools:layout="@layout/fragment_help" >
         <action
             android:id="@+id/action_HomePage_to_FirstFragment"
             app:destination="@id/FirstFragment" />
     </fragment>
     <fragment
         android:id="@+id/SecondFragment2"
-        android:name="com.example.myapplication.SecondFragment2"
+        android:name="com.example.myapplication.EndGamePage"
         android:label="Endgame"
-        tools:layout="@layout/fragment_second2" >
+        tools:layout="@layout/fragment_endgame" >
         <action
             android:id="@+id/action_SecondFragment2_to_FirstFragment"
             app:destination="@id/FirstFragment" />
